Moved homotopy-gl/gl to crate root.
